1, 클라이언트 코드 추가 [사이드바, 직원조회, 상세 정보 보기, 생성] 2. 서버 코드 수정

+++ server/src/main/java/com/ardi/server/dto/employee/request/RequestEmployeeEvaluation.java
@@ -6,7 +6,7 @@ public class RequestEmployeeEvaluation {
     public record Update(
         long idx,
         List<Long> evaluationItemIdxs,
-        Long increaseRate,
+        int increaseRate,
         List<RequestEmployeeProject.Create> projects
     ) { }
 }
